refactor: re-organize data streams v2 changes to more closely match data tracks
