+++ livekit-data-stream/src/incoming/stream_reader.rs
@@ -12,7 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::{AnyStreamInfo, ByteStreamInfo, StreamError, StreamResult, TextStreamInfo};
+use crate::info::{AnyStreamInfo, ByteStreamInfo, TextStreamInfo};
+use crate::utils::{StreamError, StreamResult};
 use bytes::{Bytes, BytesMut};
 use futures_util::{Stream, StreamExt};
 use std::{
