+++ livekit-data-stream/src/outgoing/stream_writer.rs
@@ -16,7 +16,7 @@ use std::sync::Arc;
 use tokio::sync::Mutex;
 
 use crate::info::{ByteStreamInfo, TextStreamInfo};
-use crate::outgoing::{constants::STREAM_CHUNK_SIZE_BYTES, RawStream};
+use crate::outgoing::{constants::STREAM_CHUNK_SIZE_BYTES, raw_stream::RawStream};
 use crate::utf8_chunk::Utf8AwareChunkExt;
 use crate::utils::StreamResult;
