+++ livekit/src/room/rpc/client.rs
@@ -17,7 +17,7 @@ use super::{
     ATTR_RESPONSE_TIMEOUT_MS, ATTR_VERSION, MAX_V1_PAYLOAD_BYTES, RPC_REQUEST_TOPIC,
     RPC_VERSION_V1, RPC_VERSION_V2,
 };
-use crate::data_stream::{StreamReader, StreamTextOptions, TextStreamReader};
+use crate::data_stream::api::{StreamReader, StreamTextOptions, TextStreamReader};
 use crate::room::id::ParticipantIdentity;
 use libwebrtc::native::create_random_uuid;
 use livekit_api::signal_client::CLIENT_PROTOCOL_DATA_STREAM_RPC;
